+++ libledger/DBInitializer.cpp
@@ -28,6 +28,7 @@ using namespace dev::storage;
 using namespace dev::blockverifier;
 using namespace dev::db;
 using namespace dev::eth;
+using namespace dev::mptstate;
 
 namespace dev
 {
