+++ libmptstate/CodeSizeCache.h
@@ -26,7 +26,7 @@
 
 namespace dev
 {
-namespace eth
+namespace mptstate
 {
 /**
  * @brief Simple thread-safe cache to store a mapping from code hash to code size.
