+++ libmptstate/MPTStateFactory.h
@@ -25,16 +25,16 @@
 #pragma once
 
 #include "MPTState.h"
-#include "StateFactoryInterface.h"
+#include <libexecutivecontext/StateFactoryInterface.h>
 #include <libdevcore/OverlayDB.h>
 #include <libstorage/Storage.h>
 #include <boost/filesystem.hpp>
 
 namespace dev
 {
-namespace eth
+namespace mptstate
 {
-class MPTStateFactory : public StateFactoryInterface
+class MPTStateFactory : public dev::eth::StateFactoryInterface
 {
 public:
     MPTStateFactory(u256 const& _accountStartNonce, boost::filesystem::path const& _basePath,
@@ -48,7 +48,7 @@ public:
         m_db = MPTState::openDB(m_basePath, m_genesisHash);
     };
     virtual ~MPTStateFactory(){};
-    virtual std::shared_ptr<StateFace> getState() override;
+    virtual std::shared_ptr<dev::eth::StateFace> getState() override;
 
 private:
     OverlayDB m_db;
