+++ test/unittests/libblockverifier/BlockVerifierTest.cpp
@@ -23,7 +23,7 @@
 #include <leveldb/db.h>
 #include <libblockverifier/BlockVerifier.h>
 #include <libethcore/PrecompiledContract.h>
-#include <libexecutivecontext/MPTStateFactory.h>
+#include <libmptstate/MPTStateFactory.h>
 #include <libstorage/LevelDBStorage.h>
 #include <test/tools/libutils/TestOutputHelper.h>
 #include <test/unittests/libethcore/FakeBlock.h>
@@ -33,6 +33,7 @@ using namespace dev;
 using namespace dev::eth;
 using namespace dev::storage;
 using namespace dev::blockverifier;
+using namespace dev::mptstate;
 
 namespace dev
 {
