+++ test/unittests/libexecutivecontext/ExecuteVMTest.cpp
@@ -27,7 +27,7 @@
 #include <libethcore/Transaction.h>
 #include <libexecutivecontext/Executive.h>
 #include <libexecutivecontext/ExtVM.h>
-#include <libexecutivecontext/MPTState.h>
+#include <libmptstate/MPTState.h>
 #include <test/tools/libutils/TestOutputHelper.h>
 #include <boost/test/unit_test.hpp>
 #include <iostream>
@@ -35,6 +35,7 @@
 
 using namespace std;
 using namespace dev::eth;
+using namespace dev::mptstate;
 
 namespace dev
 {
